+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ### Added
 
+- Ability to choose CSS injection method in `inject-css-in-shadow-dom` scriptlet [#477]
 - TypeScript types for CoreLibs provided [`ContentScriptApi`](./README.md#scriptlets-api--content-script-api).
 - Trusted Types API utility - [`PolicyApi`](./README.md#scriptlets-api--content-script-api--policy-api).
 
@@ -21,6 +22,7 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 - Improved docs for `json-prune`, `xml-prune` and `trusted-prune-inbound-object` scriptlets [#392]
 
+[#477]: https://github.com/AdguardTeam/Scriptlets/issues/477
 [#392]: https://github.com/AdguardTeam/Scriptlets/issues/392
 [Unreleased]: https://github.com/AdguardTeam/Scriptlets/compare/v2.1.7...HEAD